TDD: add get finance bar chart api in service

+++ src/main/kotlin/ngok3/fyp/backend/operation/event/EventEntity.kt
@@ -19,7 +19,7 @@ open class EventEntity(
     open var endDate: LocalDateTime? = null,
     open var category: String? = null,
     open var description: String? = null,
-    open var fee: Float? = null,
+    open var fee: Double? = null,
 
     ) : BaseEntity() {
     @OneToMany(mappedBy = "eventEntity")
